09-serverless/requirements.txt
onnxruntime
pillow
numpy